chatbot/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# OpenAI
openai>=1.55.0

# Vector database
qdrant-client>=1.12.0

# PostgreSQL
psycopg2-binary>=2.9.10
asyncpg>=0.30.0

# Document processing
tiktoken>=0.8.0
markdown>=3.7

# Utilities
python-dotenv>=1.0.1
httpx>=0.28.0

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
